\docType{package}
\name{pclookup-package}
\alias{pclookup}
\alias{pclookup-package}
\title{GB Postcode Centroid Lookup Table}
\source{
  \url{http://www.ons.gov.uk/ons/guide-method/geography/products/postcode-directories/-nspp-/index.html}
}
\description{
  This is a data-only package providing a look-up table of
  currently assigned postcodes in Great Britain (currently
  excluding Northern Ireland)
}
\details{
  \tabular{ll}{ Package: \tab pclookup\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2012-11-13\cr
  License: \tab GPL-2\cr }
}
\note{
  At the moment the release of this data is covered by the
  Open Government Licence, and Ordnance Survey OpenData
  Licence but the terms of the release does not cover
  Northern Ireland - hence this is a Great Britain lookup
  table and not a United Kingdon one. \cr Contains Ordnance
  Survey data (c) Crown copyright and database right 2012
  \cr Contains National Statistics data (c) Crown copyright
  and database right 2012
}
\author{
  Chris Brunsdon <brunsdon@liverpool.ac.uk> \cr
}
\keyword{package}

