\docType{data}
\name{postcodes}
\alias{postcodes}
\title{GB Postcode Lookup Table}
\format{A data frame with 1746831 observations on the following 5 variables.
\describe{ \item{\code{postcode}}{A GB unit postcode}
\item{\code{largeuser}}{Boolean indicator of whether postcode is for a large
user} \item{\code{easting}}{Easting of grid reference for postcode centroid}
\item{\code{northing}}{Northing of grid reference for postcode centroid}
\item{\code{quality}}{a numeric quantity indicating the quality of the
centroid - see ONS documentation via web link }}}
\source{
  \url{http://www.ons.gov.uk/ons/guide-method/geography/products/postcode-directories/-nspp-/index.html}
}
\description{
  A data frame containing Ordnance Survey National Grid
  coordinates of postcode units.
}
\keyword{datasets}

