\docType{data}
\name{proj4.used}
\alias{proj4.used}
\title{Proj4 string for grid coordinates}
\source{
  \url{http://spatialreference.org/ref/sr-org/6971/}
}
\description{
  A character variable containing the Proj4 string used to
  represent the National Grid projection used for
  coordinates
}
\keyword{datasets}

