\docType{data}
\name{release.date}
\alias{release.date}
\title{Proj4 string for grid coordinates}
\source{
  \url{http://www.ons.gov.uk/ons/guide-method/geography/products/postcode-directories/-nspp-/index.html}
}
\description{
  A character variable containing the relaese date of this
  version of the postcode location data
}
\keyword{datasets}

